build refactoring: add -Wextra (and -Wno-unused-parameter)
This ensures a bit more coding strictness. -Wno-unused-parameter is
and exception, because some ds/filters/outputs do not use argument
parameter, but for the sake of unified provider prototypes (and their
pointers in common array) they must be identical in this respect.

+++ src/datasource/egroup.c
@@ -55,7 +55,7 @@ int snoopy_datasource_egroup (char *result, char *arg)
     struct group   gr;
     struct group  *gr_gid         = NULL;
     char          *buffgr_gid     = NULL;
-    size_t         buffgrsize_gid = 0;
+    long           buffgrsize_gid = 0;
     int            messageLength  = 0;
 
     /* Allocate memory */
