+++ src/datasource/tty_username.c
@@ -62,7 +62,7 @@ int snoopy_datasource_tty_username (char *result, char *arg)
     struct passwd  pwd;
     struct passwd *pwd_uid = NULL;
     char          *buffpwd_uid = NULL;
-    size_t         buffpwdsize_uid = 0;
+    long           buffpwdsize_uid = 0;
     int            messageLength   = 0;
 
     /* Get tty path */
